+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed `Request devicetree/customActions failed with message: Cannot read properties of undefined (reading 'uri")`
   in VSCode when using context menu document to `Copy C identifier`
+- Fixed formatting files when file has setting that differ from workspace default.
 
 ## [0.9.4] - 2026-05-22